Excluded Javadoc reference checking in test code
If the check is on, a reference to a class or method in a Javadoc of the
class which cannot access the referenced class or method, makes this
reference an error.

In some cases we want to reference classes that are not accessible
from test environment classes to improve navigation from the test
environment code to the code of tests that use these test environment
classes.