Remove `LicenseReport` dependency object
We still use License Report, but dependency on it is declared in `buildSrc/build.gradle.kts`.

There is no way to import a Kotlin object in this file, because it's the very beginning of the compilation process, and the object is not available yet. Previously, the object was used in the Groovy script plugins, which are no longer used.